fix pull request conflicts

+++ src/Infra/Interfaces/IUsersRepository.cs
@@ -4,7 +4,7 @@ namespace Infra.Interfaces;
 
 public interface IUsersRepository
 {
-    Task<UserDto?> GetUserById(int id);
+    Task<UserDto?> GetUserById(ulong id);
     Task<UserDto?> GetUserByPhone(string phone, bool withRoleUsers = false);
     Task<UserDto?> GetUserByAuthId(string authId, bool withRoleUsers = false);
 }
